requirements.txt
alembic==1.17.2
babel==2.17.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
docxcompose==1.4.0
docxtpl==0.18.0
Flask==3.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
python-docx==1.1.2
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
